Labfiles/07-agent-observability/Python/requirements.txt
azure-ai-projects
azure-ai-agents
azure-ai-evaluation
azure-identity
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
langgraph
langchain_openai
langchain_community
python-dotenv